requirements.txt
boto3==1.4.5

Flask==0.12.2

Pillow==4.2.1

pymongo==3.4.0